Shared classes updated.

+++ TextToSpeech/JocysCom/Drawing/Effects/Transparent.cs
@@ -22,7 +22,7 @@ namespace JocysCom.ClassLibrary.Drawing
 				for (int x = 0; x < w; x++)
 				{
 					p = b.GetPixel(x, y);
-					a = (int)((float)p.A * (float)alpha / byte.MaxValue);
+					a = (int)(p.A * (float)alpha / byte.MaxValue);
 					if (a >= byte.MaxValue) a = byte.MaxValue;
 					if (a <= byte.MinValue) a = byte.MinValue;
 					b.SetPixel(x, y, Color.FromArgb(a, p.R, p.G, p.B));
